+++ rmad/utils/cbi_utils.h
@@ -15,7 +15,7 @@ class CbiUtils {
   virtual ~CbiUtils() = default;
 
   // Get the sku id of the device from cbi.
-  virtual bool GetSkuId(uint64_t* sku_id) const = 0;
+  virtual bool GetSkuId(uint32_t* sku_id) const = 0;
 
   // Get the dram part number of the device from cbi.
   virtual bool GetDramPartNum(std::string* dram_part_num) const = 0;
@@ -24,7 +24,7 @@ class CbiUtils {
   virtual bool GetSsfc(uint32_t* ssfc) const = 0;
 
   // Set the sku id of the device to cbi.
-  virtual bool SetSkuId(uint64_t sku_id) = 0;
+  virtual bool SetSkuId(uint32_t sku_id) = 0;
 
   // Set the dram part number of the device to cbi.
   virtual bool SetDramPartNum(const std::string& dram_part_num) = 0;
